Add files via upload
ntroduction
In this project you will implement a program which works with 2D figures.
Initially, only a limited set of figures will be available to the end user (they are listed
below). However, in the future this may change. It should be easy to add support for new
figures to the program or to remove existing ones.
• Triangle — has three sides, specified by their length.
• Circle — has a radius.
• Rectangle — has two sides, specified by their lengths.
For all figures it should be possible to calculate their perimeter.
Figures must be immutable. It is not necessary to be able to access their specific proper
ties (i.e. the sides of a triangle, a circle’s radius, etc.). It is enough to be able to calculate
their perimeter.
All figures can be represented as a string. This is done by specifying the figure type and
then listing its parameters, using the space character as separator. For example:
• "triangle 10 20 30" — a triangle with sides 10, 20, and 30.
• "circle 2.2" — a circle with a radius of 2.2.
• etc.
Your solution should allow the user to create a figure from its string representation, or
to convert a figure to a string.
When the program starts, it asks the user to choose how they prefer to create new fig
ures. All of the methods listed below must be available. Also, it should be easy to add new
methods to the program.
• Random — figures are created using a random number generator.
• STDIN — figures are read from STDIN.
• File — figures are read from a text file specified by the user.
For the last two options you don’t have to invent new syntax for entering figures. Instead,
you can use the standard string representation for figures.
When a method of entering figures has been chosen, the user should be able to enter a
list of figures. After the figures have been entered, the user should be able to:
• List all figures to STDOUT.
• Delete a figure from the list.
• Duplicate a figure in the list. The copy should be added to the end of the list.
• Store the resulting list back into a file.